users/requirements.txt
Django==4.2
psycopg2-binary==2.9.9
Faker==26.1.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
Pillow
django-cors-headers==4.4.0